requirements.txt
ollama
python-dotenv
langchain
langchain-ollama
